+++ tests/impls/yolo/test_yolo_run.py
@@ -52,7 +52,6 @@ def yolo_run(version: int) -> None:
 
     engine = trtutils.impls.yolo.YOLO(
         engine_path,
-        version=version,
         warmup=False,
     )
 
@@ -67,7 +66,7 @@ def multiple_yolos_run(version: int) -> None:
     engine_path = build_yolo(version)
 
     engines = [
-        trtutils.impls.yolo.YOLO(engine_path, version=version, warmup=False) for _ in range(4)
+        trtutils.impls.yolo.YOLO(engine_path, warmup=False) for _ in range(4)
     ]
 
     outputs = [engine.mock_run() for engine in engines]
@@ -87,7 +86,6 @@ def yolo_run_in_thread(version: int) -> None:
 
         engine = trtutils.impls.yolo.YOLO(
             engine_path,
-            version=7,
             warmup=False,
         )
 
@@ -117,7 +115,6 @@ def multiple_yolos_run_in_threads(version: int) -> None:
 
         engine = trtutils.impls.yolo.YOLO(
             engine_path,
-            version=version,
             warmup=False,
         )